Put everything in a single module according to review notes.
