--- architecture-hexagon/domain/.gitignore
@@ -1,4 +1,0 @@
-*.iml
-*.ipr
-*.iws
-target/
