--- architecture-hexagon/domain/src/main/java/com/baeldung/hexagon/tour/suggest/api/TourSuggestApi.java
@@ -1,9 +1,0 @@
-package com.baeldung.hexagon.tour.suggest.api;
-
-import com.baeldung.hexagon.tour.suggest.dto.Tour;
-
-import java.util.List;
-
-public interface TourSuggestApi {
-    List<Tour> suggestBestPrice(String origin, String destination, int count);
-}
