--- architecture-hexagon/domain/src/main/java/com/baeldung/hexagon/tour/suggest/spi/TourFetchSpi.java
@@ -1,9 +1,0 @@
-package com.baeldung.hexagon.tour.suggest.spi;
-
-import com.baeldung.hexagon.tour.suggest.dto.Tour;
-
-import java.util.List;
-
-public interface TourFetchSpi {
-    List<Tour> fetchToursFromTo(String origin, String destination);
-}
