--- architecture-hexagon/infrastructure/repository-inmemory/.gitignore
@@ -1,4 +1,0 @@
-*.iml
-*.ipr
-*.iws
-target/
