--- architecture-hexagon/infrastructure/repository-inmemory/src/main/java/com/baeldung/hexagon/tour/suggest/model/TourModel.java
@@ -1,31 +1,0 @@
-package com.baeldung.hexagon.tour.suggest.model;
-
-public class TourModel {
-    private String source;
-    private String destination;
-    private String name;
-    private int price;
-
-    public TourModel(String source, String destination, String name, int price) {
-        this.source = source;
-        this.destination = destination;
-        this.name = name;
-        this.price = price;
-    }
-
-    public String getSource() {
-        return source;
-    }
-
-    public String getDestination() {
-        return destination;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-}
