--- architecture-hexagon/infrastructure/sms-adapter/.gitignore
@@ -1,4 +1,0 @@
-*.iml
-*.ipr
-*.iws
-target/
