--- architecture-hexagon/infrastructure/sms-application/.gitignore
@@ -1,4 +1,0 @@
-*.iml
-*.ipr
-*.iws
-target/
