--- architecture-hexagon/infrastructure/sms-application/pom.xml
@@ -1,49 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>sms-application</artifactId>
-    <name>sms-application</name>
-    <packaging>jar</packaging>
-
-    <parent>
-        <groupId>com.baeldung.architecture-hexagonal</groupId>
-        <artifactId>infrastructure</artifactId>
-        <version>0.0.1-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.baeldung.architecture-hexagonal</groupId>
-            <artifactId>domain</artifactId>
-            <version>${tour.project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.baeldung.architecture-hexagonal</groupId>
-            <artifactId>repository-inmemory</artifactId>
-            <version>${tour.project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.baeldung.architecture-hexagonal</groupId>
-            <artifactId>sms-adapter</artifactId>
-            <version>${tour.project.version}</version>
-        </dependency>
-    </dependencies>
-
-
-    <build>
-        <pluginManagement>
-            <plugins>
-                <plugin>
-                    <groupId>org.apache.maven.plugins</groupId>
-                    <artifactId>maven-compiler-plugin</artifactId>
-                    <version>3.5.1</version>
-                    <configuration>
-                        <source>1.8</source>
-                        <target>1.8</target>
-                    </configuration>
-                </plugin>
-            </plugins>
-        </pluginManagement>
-    </build>
-</project>
