+++ architecture-hexagon/src/main/java/com/baeldung/hexagonal/tour/domain/suggest/dto/Tour.java
@@ -1,0 +1,19 @@
+package com.baeldung.hexagonal.tour.domain.suggest.dto;
+
+public class Tour {
+    private String name;
+    private int price;
+
+    public Tour(String name, int price) {
+        this.name = name;
+        this.price = price;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getPrice() {
+        return price;
+    }
+}
